Local PC
Space Invaders: Final Frontier" is a Python-based 2D arcade shooter game, developed as a comprehensive university project. This game serves as a modern reinterpretation of the classic "Space Invaders," built utilizing the Pygame library. The core objective is to offer an engaging and progressively challenging experience, where players control a spaceship to defend against waves of enemy UFOs, aiming for survival and high scores.